Address comments from PR
