+++ hdijupyterutils/hdijupyterutils/utils.py
@@ -3,10 +3,6 @@
 import os
 import uuid
 
-import numpy as np
-import pandas as pd
-
-from .filesystemreaderwriter import FileSystemReaderWriter
 
 first_run = True
 instance_id = None
@@ -35,22 +31,3 @@ def get_instance_id():
         raise ValueError("Tried to return empty instance ID.")
 
     return instance_id
-
-
-def coerce_pandas_df_to_numeric_datetime(df):
-    for column_name in df.columns:
-        coerced = False
-
-        if not coerced and df[column_name].dtype == np.dtype("object"):
-            try:
-                df[column_name] = pd.to_datetime(df[column_name], errors="raise")
-                coerced = True
-            except (ValueError, TypeError, OverflowError):
-                pass
-
-        if not coerced and df[column_name].dtype == np.dtype("object"):
-            try:
-                df[column_name] = pd.to_numeric(df[column_name], errors="raise")
-                coerced = True
-            except (ValueError, TypeError):
-                pass
